requirements.txt
mysqlclient==1.4.4
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
sqlalchemy-migrate==0.12.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
bcrypt==3.1.7
python-dotenv==0.10.3
python-ldap==3.2.0
Flask-Admin==1.5.3
six==1.12.0
zeep==3.4.0
requests==2.22.0
openpyxl==3.0.0
python-dateutil==2.8.0
celery==4.3.0